requirement.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-barcode[images]==0.14.0
Pillow==10.1.0
pydantic==2.5.0
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
